Add support for packages without versions in name
This checks if the package is named gz-cmake or depends
on gz-cmake to determine if it should have a version
in its package name. Some logic for getting package
version numbers is changed to get it directly from
the package.xml file using the split_version helper
since it may not be available from the package name.
Extra cmake config files are not written or installed
for packages with unversioned names.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>