Jetty: support package names without version number (#19)
This checks if the package is named gz-cmake or depends
on gz-cmake to determine if it should have a version
in its package name. Some logic for getting package
version numbers is changed to get it directly from
the package.xml file using the split_version helper
since it may not be available from the package name.
For packages with unversioned names, bybind11 is not
skipped, and extra cmake config files are not written or
installed .

Some additional clean up of helper functions:
* Rename is_gz_cmake4 to is_gz_cmake4_or_later to match
  its documentated purpose. Add logic to recognize
  gz-cmake without version number as later than gz-cmake4.
* Change parameter name to pkg_major_version
* Fix documentation of VENDOR_FROM_LIB_VCS_REF

Signed-off-by: Steve Peters <scpeters@openrobotics.org>