Overwrite CMake configs by default
The default was set to False because we had made a manual change in
`gz_cmake_vendor` that we didn't want to overwrite. However, this
causes problems when we introduce new config files. For example, we
recently added a `.in` files to configure PYTHONPATH for each vendor
package. Since the default is False, when the vendor package is created
by `release-tools/release.py`, it doesn't include the new config files.

The solution here to programmatically achieve what was done manually in
gz_cmake_vendor and set `overwrite_cmake_configs` to True by default.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>